Fix albedo issues
If user asks for outputs at a temporal resolution of >daily (>86400s),
the program first averages the data to this resolution prior to
calculating the heat fluxes. In this case, the hour component will be
changed to zero (default in matlab). Thus, the program will calculate
the albedo at midnight (=1), resulting in all of the short-wave
radiation being reflected. This fix calculates the albedo at midday for
the daily data. This is an estimate of the daily average albedo for a
given day.